refactor: rename population/base.py to population/population.py
